hub/lora_rfm95.py
# lora_rfm95.py - Forest Guardian Hub
import spidev
import RPi.GPIO as GPIO
import time
import logging

LORA_SCK = 11
LORA_MISO = 9
LORA_MOSI = 10
LORA_NSS = 8
LORA_RST = 25
LORA_DIO0 = 24
LORA_FREQ = 915e6

class RFM95:
    def __init__(self):
        GPIO.setmode(GPIO.BCM)
        GPIO.setup(LORA_RST, GPIO.OUT)
        GPIO.setup(LORA_NSS, GPIO.OUT)
        GPIO.setup(LORA_DIO0, GPIO.IN)
        self.spi = spidev.SpiDev()
        self.spi.open(0, 0)
        self.spi.max_speed_hz = 5000000
        self.reset()

    def reset(self):
        GPIO.output(LORA_RST, GPIO.LOW)
        time.sleep(0.1)
        GPIO.output(LORA_RST, GPIO.HIGH)
        time.sleep(0.1)

    def read_packet(self):
        # Placeholder: Implement RFM95W receive logic
        # For demo, return None
        return None

    def close(self):
        self.spi.close()
        GPIO.cleanup()

lora = RFM95()
