hub/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.1
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.29
PyJWT==2.8.0
requests==2.31.0
azure-iot-device==2.11.1
openai==1.30.1
azure-identity==1.16.0
azure-cosmos==4.6.0
geopy==2.4.1
spidev==3.6
RPi.GPIO==0.7.1
